simpler implementation that just sends a single message per sendViaWebSocket
