+++ packages/expo-dev-launcher/android/src/main/java/expo/modules/devlauncher/logs/DevLauncherRemoteLog.kt
@@ -5,22 +5,16 @@ import com.google.gson.annotations.Expose
 
 @Suppress("UNUSED")
 internal data class DevLauncherRemoteLog(
-  val logBody: String,
+  val messages: List<String>,
   @Expose val level: String = "error",
   @Expose val mode: String = "BRIDGE"
 ) {
-  constructor(
-    throwable: Throwable,
-    level: String = "error",
-    mode: String = "BRIDGE"
-  ) : this(throwable.toRemoteLogString(), level, mode)
-
   /**
-   * `data` is an array whose members are simply concatenated before printing, so we use a trivial
-   * array of just a single string (which may span multiple lines).
+   * `data` is an array whose members are simply concatenated with a space before printing to the
+   * console, so we join messages with a newline and send an array consisting of just a single item.
    */
   @Expose
-  private val data = arrayOf(logBody)
+  private val data = arrayOf(messages.joinToString("\n"))
 
   @Expose
   private val type = "log"
@@ -31,18 +25,4 @@ internal data class DevLauncherRemoteLog(
       .create()
       .toJson(this)
   }
-}
-
-internal fun Throwable.toRemoteLogString(): String {
-  val separator = "\n  "
-  val baseTrace = stackTrace.joinToString(separator) {
-    it.toString()
-  }
-  val remoteLogString = "$this$separator$baseTrace"
-
-  cause?.let {
-    return "$remoteLogString\nCaused by ${it.toRemoteLogString()}"
-  }
-
-  return remoteLogString
 }
